pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tsu-cli"
version = "0.1.0"
description = "Generate project tech documentation and push to Confluence using GitHub Copilot SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    # { name = "Vivek Rana", email = "vivekrana.2012@gmail.com" },
]
keywords = ["documentation", "confluence", "copilot", "cli", "markdown"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "github-copilot-sdk>=0.1.0",
    "typer[all]>=0.12.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "keyring>=25.0.0",
    "markdown>=3.5.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
tsu = "tsu_cli.main:app"

# [project.urls]
# Homepage = "https://github.com/vivekrana2012/tsu-cli"
# Repository = "https://github.com/vivekrana2012/tsu-cli.git"
# Issues = "https://github.com/vivekrana2012/tsu-cli/issues"

[tool.setuptools]
packages = ["tsu_cli", "tsu_cli.prompts"]

[tool.setuptools.package-dir]
tsu_cli = "src"
"tsu_cli.prompts" = "src/prompts"

[tool.setuptools.package-data]
tsu_cli = ["*.md"]
"tsu_cli.prompts" = ["*.md"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
